move CodeQuality to root Rector
