+++ src/Rector/CodeQuality/InArrayAndArrayKeysToArrayKeyExistsRector.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Rector\Contrib\CodeQuality;
+namespace Rector\Rector\CodeQuality;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\FuncCall;
