+++ tests/Rector/CodeQuality/InArrayAndArrayKeysToArrayKeyExistsRector/Correct/correct.php.inc
