+++ tests/Rector/CodeQuality/InArrayAndArrayKeysToArrayKeyExistsRector/config.yml
@@ -1,3 +1,2 @@
 services:
     Rector\Rector\CodeQuality\InArrayAndArrayKeysToArrayKeyExistsRector: ~
-    Rector\Rector\CodeQuality\UnnecessaryTernaryExpressionRector: ~
