+++ tests/Rector/CodeQuality/UnnecessaryTernaryExpressionRector/Correct/correct.php.inc
