+++ tests/Rector/CodeQuality/UnnecessaryTernaryExpressionRector/Wrong/wrong.php.inc
