--- tests/Rector/Contrib/CodeQuality/InArrayAndArrayKeysToArrayKeyExistsRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\CodeQuality\InArrayAndArrayKeysToArrayKeyExistsRector: ~
