--- tests/Rector/Contrib/CodeQuality/UnnecessaryTernaryExpressionRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\CodeQuality\UnnecessaryTernaryExpressionRector: ~
